requirements.txt
cmake==3.17.2
dacite==1.5.0
jupyterlab==2.1.2
matplotlib==3.2.1
MulticoreTSNE==0.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pyarrow==0.17.0
pydot==1.4.1
pylint==2.5.2
scipy==1.4.1
scikit-learn==0.23.1
tensorflow-gpu==2.2.0
tensorflow-addons==0.10.0
tqdm==4.46.0
